+++ apps/hubble/src/storage/stores/storeEventHandler.ts
@@ -20,15 +20,17 @@ import {
 import AsyncLock from 'async-lock';
 import { err, ok, ResultAsync } from 'neverthrow';
 import { TypedEmitter } from 'tiny-typed-emitter';
-import RocksDB, { Iterator, Transaction } from '~/storage/db/rocksdb';
-import { RootPrefix, UserMessagePostfix } from '~/storage/db/types';
-import { StorageCache } from '~/storage/stores/storageCache';
-import { makeTsHash } from '~/storage/db/message';
+import RocksDB, { Iterator, Transaction } from '../db/rocksdb.js';
+import { RootPrefix, UserMessagePostfix } from '../db/types.js';
+import { StorageCache } from './storageCache.js';
+import { makeTsHash } from '../db/message.js';
 import {
   bytesCompare,
   CastAddMessage,
   CastRemoveMessage,
   getFarcasterTime,
+  LinkAddMessage,
+  LinkRemoveMessage,
   ReactionAddMessage,
   ReactionRemoveMessage,
   SignerAddMessage,
@@ -51,7 +53,9 @@ type PrunableMessage =
   | SignerRemoveMessage
   | UserDataAddMessage
   | VerificationAddEthAddressMessage
-  | VerificationRemoveMessage;
+  | VerificationRemoveMessage
+  | LinkAddMessage
+  | LinkRemoveMessage;
 
 export type StoreEvents = {
   /**
@@ -110,14 +114,14 @@ export class HubEventIdGenerator {
   private _lastSeq: number;
   private _epoch: number;
 
-  constructor(options: { epoch?: number; lastTimestamp?: 0; lastIndex?: 0 } = {}) {
+  constructor(options: { epoch?: number; lastTimestamp?: number; lastIndex?: number } = {}) {
     this._epoch = options.epoch ?? 0;
     this._lastTimestamp = options.lastTimestamp ?? 0;
     this._lastSeq = options.lastIndex ?? 0;
   }
 
-  generateId(): HubResult<number> {
-    const timestamp = Date.now() - this._epoch;
+  generateId(options: { currentTimestamp?: number } = {}): HubResult<number> {
+    const timestamp = (options.currentTimestamp || Date.now()) - this._epoch;
 
     if (timestamp === this._lastTimestamp) {
       this._lastSeq = this._lastSeq + 1;
@@ -126,11 +130,11 @@ export class HubEventIdGenerator {
       this._lastSeq = 0;
     }
 
-    if (this._lastTimestamp > 2 ** TIMESTAMP_BITS) {
+    if (this._lastTimestamp >= 2 ** TIMESTAMP_BITS) {
       return err(new HubError('bad_request.invalid_param', `timestamp > ${TIMESTAMP_BITS} bits`));
     }
 
-    if (this._lastSeq > 2 ** SEQUENCE_BITS) {
+    if (this._lastSeq >= 2 ** SEQUENCE_BITS) {
       return err(new HubError('bad_request.invalid_param', `sequence > ${SEQUENCE_BITS} bits`));
     }