+++ apps/hubble/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @farcaster/hubble
 
+## 1.3.0
+
+### Minor Changes
+
+- 2bc598a: Added support for Links
+
+### Patch Changes
+
+- 37f515f: Add app version to contactInfo message
+- 7f050d2: Correctly error when sequence ID exceeds max allowed value
+- 27181c6: added fallback to dbreset
+- 16ea9b4: Validate the eth network in ethEventsProvider
+- Updated dependencies [2bc598a]
+  - @farcaster/hub-nodejs@0.8.0
+
+## 1.2.4
+
+### Patch Changes
+
+- Re-publish
+- Updated dependencies
+  - @farcaster/hub-nodejs@0.7.4
+
+## 1.2.3
+
+### Patch Changes
+
+- c4891754: Add --watch to hub status command and more readable output
+- 2ca66b17: replace @noble/ed25519 with faster and more secure @noble/curves
+- 43444471: fix(eth): add stronger retry logic to eth events provider
+- ea55abcb: Include semver as explicit dependency
+- fe755fbd: Switch from `rocksdb` to `@farcaster/rocksdb` NPM package
+- 651df412: Add @faker-js/faker as prod dependency
+- Updated dependencies [2ca66b17]
+  - @farcaster/hub-nodejs@0.7.3
+
 ## 1.2.2
 
 ### Patch Changes